+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.5] - 2024-07-18
+
 ### Added
 
 - Search filter for link translations.
@@ -408,7 +410,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First Release!
 
-[unreleased]: https://github.com/26b/unbabble/compare/0.4.4...HEAD
+[unreleased]: https://github.com/26b/unbabble/compare/0.4.5...HEAD
+[0.4.5]: https://github.com/26b/unbabble/compare/0.4.4...0.4.5
 [0.4.4]: https://github.com/26b/unbabble/compare/0.4.3...0.4.4
 [0.4.3]: https://github.com/26b/unbabble/compare/0.4.2...0.4.3
 [0.4.2]: https://github.com/26b/unbabble/compare/0.4.1...0.4.2